product to float, add Tomasz

+++ src/main/java/com/example/demo/Product.java
@@ -12,7 +12,7 @@ import java.util.HashMap;
 public class Product {
 
     private String name;
-    private HashMap<Integer,Float> quantityPriceMap;
+    private HashMap<Float,Float> quantityPriceMap;
     private Float myPrice;
 
 }
